Function chain_path_derivation::bip44::new

pub fn new() -> DerivationPath<Bip44<Root>>

create a Bip44 chain path derivation

This derivation level is not really interesting, though it is interesting to consider the following levels which can then be constructed via each individual type.

See module documentation for more details