Crate chain_path_derivation

BIP44 addressing

provides all the logic to create safe sequential addresses using BIP44 specification.

Modules

bip44

Bip44 derivation scheme

rindex

Random Index scheme

Structs

AnyScheme

any derivation path scheme

Derivation

a derivation value that can be used to derive keys

DerivationPath

a derivation path with a tagged scheme (for example Bip44).

DerivationPathRange

DerivationRange

iterator to create derivation values

HardDerivation

wrapper to guarantee the given derivation is a soft derivation

HardDerivationRange

iterator to create derivation values

SoftDerivation

wrapper to guarantee the given derivation is a soft derivation

SoftDerivationRange

iterator to create derivation values

Enums

DerivationError

ParseDerivationError

ParseDerivationPathError